Rename setup_function to db_setup_functions.
